refactor: split gocoveragecheck evaluation
